chore: move cancel function to cancel dialog
